Update Dockerfile to build production image
