+++ laa_access_civil_legal_aid/app/__init__.py
@@ -8,7 +8,7 @@ from flask_wtf.csrf import CSRFProtect
 from govuk_frontend_wtf.main import WTFormsHelpers
 from jinja2 import ChoiceLoader, PackageLoader, PrefixLoader
 
-from config import Config
+from app.config import Config
 
 compress = Compress()
 csrf = CSRFProtect()
@@ -103,12 +103,12 @@ def create_app(config_class=Config):
     compress.init_app(app)
     csrf.init_app(app)
     limiter.init_app(app)
-    talisman.init_app(app, 
-                      content_security_policy=csp, 
-                      permissions_policy=permissions_policy, 
-                      content_security_policy_nonce_in=["script-src"], 
-                      force_https=True,
-                      session_cookie_secure=True,
+    talisman.init_app(app,
+                      content_security_policy=csp,
+                      permissions_policy=permissions_policy,
+                      content_security_policy_nonce_in=["script-src"],
+                      force_https=False,
+                      session_cookie_secure=False,
                       session_cookie_http_only=True,
                       session_cookie_samesite="Strict",)
