+++ laa_access_civil_legal_aid/compose.yml
@@ -15,5 +15,5 @@ services:
     ports:
       - "8000:8000"
     volumes:
-      - .:/home/containeruser
+      - .:/home/app/access
