+++ laa_access_civil_legal_aid/requirements/source/requirements_base.in
@@ -1,0 +1,14 @@
+Flask
+Flask-Assets
+Flask-Compress
+Flask-Limiter
+flask-talisman
+Flask-WTF
+govuk-frontend-wtf
+click
+python-dotenv
+requests
+cssmin  # used by Flask-Assets
+jsmin  # used by Flask-Assets
+libsass # used by Flask-Assets
+PyYAML
