+++ laa_access_civil_legal_aid/requirements/source/requirements_development.in
@@ -1,3 +1,4 @@
+-r requirements_base.in
 black
 flake8-bugbear
 isort
